\name{flowWorkspaceDataInfo}
\alias{flowWorkspaceDataInfo}
\title{
Print some information about this package and the data.
}
\description{
Contains the data necessary to build the flowWorkspace vignette.
}
\usage{
flowWorkspaceDataInfo()
}
\author{
Greg Finak
}


\seealso{
\code{\link[flowWorkspace]{flowWorkspace-package}}
}
\examples{
flowWorkspaceDataInfo();
}
