\name{flowWorkspaceData-package}
\alias{flowWorkspaceData-package}
\alias{flowWorkspaceData}
\docType{package}
\title{
A flowJo XML Workspace and Associated Data for testing the flowWorkspace package.
}
\description{
This package contains the data necessary to build the flowWorkspace vignette. The primary purpose of this package is to test the import and export code in the flowWorkspace package.
}
\details{
\tabular{ll}{
Package: \tab flowWorkspaceData\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2011-03-31\cr
License: \tab GPL 2.0\cr
LazyLoad: \tab yes\cr
}
flowWorkspaceDataInfo() provides information about the package.
}
\author{
Greg Finak

Maintainer: <gfinak@fhcrc.org>
}
\keyword{ package }
\seealso{
\code{\link[flowWorkspace:flowWorkspace-package]{flowWorkspace}}
}
\examples{
flowWorkspaceDataInfo()
}
